+++ src/main/java/kg/attractor/xfood/repository/CheckTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CheckTypeRepository extends JpaRepository<CheckType, Integer> {
+public interface CheckTypeRepository extends JpaRepository<CheckType, Long> {
 }
