+++ src/main/java/kg/attractor/xfood/repository/PizzeriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PizzeriaRepository extends JpaRepository<Pizzeria, Integer> {
+public interface PizzeriaRepository extends JpaRepository<Pizzeria, Long> {
 }
